Ansible Provisioning YAML
Ansible Provisioning YAML
This Ansible playbook (provision.yaml) is designed to pull a Docker image for deployment on target servers. It includes the following tasks:

Name: The playbook includes a single task named "Pull the Docker image," which describes the main objective of the playbook.

Hosts: The playbook targets servers labeled as web-servers for the Docker image pull operation.

Become: The playbook utilizes elevated privileges (become: yes) to execute tasks as a superuser or with appropriate permissions.

Task: The task within the playbook employs the Ansible docker_image module to manage Docker images.

Docker Image Name: Specifies the name of the Docker image to pull (docker.io/ubuntu123am/my-apache-image).

Source: Indicates that the image should be pulled from a specified source (source: pull).

State: Ensures that the image is present on the target servers (state: present).

This Ansible playbook is a crucial part of the overall deployment process, ensuring that the required Docker image is available on the designated servers for subsequent deployment steps.